requirements.txt
# Government Policy Intelligence Platform - Requirements
# Python 3.11+ recommended

# ============================================
# LangChain & LLM Integrations
# ============================================
langchain==1.1.0
langchain-core==1.1.0
langgraph==1.0.4
langchain-openai==1.1.0
openai==2.8.1
langchain-anthropic==1.2.0
langchain-google-genai==0.0.1
google-generativeai==0.3.2
langchain-huggingface==1.1.0
transformers==4.57.3
huggingface-hub==0.36.0

# ============================================
# Web Framework & API
# ============================================
fastapi==0.115.12
starlette==0.46.1
uvicorn==0.34.0
httpx==0.28.1
httpcore==1.0.7
h11==0.14.0
httptools==0.6.4
watchfiles==1.0.4
typer==0.15.2
python-multipart==0.0.20

# ============================================
# Database & ORM
# ============================================
SQLAlchemy==1.4.0
alembic==1.15.2
psycopg2-binary==2.9.10

# ============================================
# Supabase Client & Dependencies
# ============================================
supabase==2.11.0
gotrue==2.12.4
postgrest==0.19.3
realtime==2.24.0
storage3==0.11.3
supafunc==0.9.4

# ============================================
# Document Processing & OCR
# ============================================
PyMuPDF==1.25.2
python-docx==1.1.2
Pillow==11.1.0
easyocr==1.7.2
opencv-python-headless==4.12.0.88
scikit-image==0.25.2
python-bidi==0.6.7
lxml==6.0.2

# ============================================
# Vector Embeddings & Search
# ============================================
faiss-cpu==1.9.0
sentence-transformers==3.3.1
# Note: For GPU support, install PyTorch with CUDA:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# ============================================
# Authentication & Security
# ============================================
python-jose==3.4.0
bcrypt==4.3.0
cryptography==44.0.2
passlib==1.7.4
PyJWT==2.10.1

# ============================================
# Data Validation & Serialization
# ============================================
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.8.1
pydantic-extra-types==2.10.3
orjson==3.10.16
ujson==5.10.0

# ============================================
# Utilities & Core Dependencies
# ============================================
python-dotenv==1.1.0
numpy==2.2.4
scikit-learn==1.7.2
scipy==1.16.3
PyYAML==6.0.2
click==8.1.8
colorama==0.4.6
rich==13.9.4
typing_extensions==4.14.0
python-dateutil==2.9.0.post0

# ============================================
# Async & HTTP
# ============================================
aiohttp==3.13.2
asyncio==3.4.3
h2==4.3.0
httpx[http2]==0.28.1

# ============================================
# Additional Dependencies
# ============================================
email_validator==2.2.0
Jinja2==3.1.6
Mako==1.3.9
sniffio==1.3.1
websockets==15.0.1
markdown-it-py==3.0.0
Pygments==2.19.1
Shapely==2.1.2
pyclipper==1.3.0.post6
ninja==1.13.0

# ============================================
# Development & Testing
# ============================================
pytest==9.0.1
pytest-mock==3.15.1
